[WFLY-15371] Move the class config source test to own package
