+++ testsuite/integration/microprofile/src/test/java/org/wildfly/test/integration/microprofile/config/smallrye/management/config_source/from_class/CustomConfigSourceServiceLoader.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.wildfly.test.integration.microprofile.config.smallrye.management.config_source;
+package org.wildfly.test.integration.microprofile.config.smallrye.management.config_source.from_class;
 
 import java.util.Collections;
 import java.util.HashMap;
@@ -31,17 +31,24 @@ import org.eclipse.microprofile.config.spi.ConfigSource;
 /**
  * @author <a href="mailto:mjurc@redhat.com">Michal Jurc</a> (c) 2018 Red Hat, Inc.
  */
-public class CustomConfigSourceAServiceLoader implements ConfigSource {
+public class CustomConfigSourceServiceLoader implements ConfigSource {
 
+    public static final String PROP_NAME = "my.prop.from.service.loader.class";
+    public static final String PROP_VALUE = "I'm from a custom config source provided by service loader mechanism!";
+    public static final String PROP_NAME_OVERRIDEN_BY_SERVICE_LOADER = "my.prop.from.class.overriden.service.loader";
+    public static final String PROP_VALUE_OVERRIDEN_BY_SERVICE_LOADER = "I'm from a custom config source provided by service" +
+            " loader mechanism! My ordinality is higher, so I override property from custom config source without service loader.";
     public static final String PROP_NAME_SAME_ORDINALITY_OVERRIDE = "my.prop.from.class.overriden.same.ordinality";
     public static final String PROP_VALUE_SAME_ORDINALITY_OVERRIDE = "I'm from a custom config source provided by service " +
-            "loader overriding ConfigSource with same ordinality since my FQCN is ranked higher lexicographically.";
-
+            "loader! However I should be overriden by property from ConfigSource with same ordinality based on FQCN " +
+            "lexicographic ordering.";
 
     final Map<String, String> props;
 
-    public CustomConfigSourceAServiceLoader() {
+    public CustomConfigSourceServiceLoader() {
         props = new HashMap<>();
+        props.put(PROP_NAME, PROP_VALUE);
+        props.put(PROP_NAME_OVERRIDEN_BY_SERVICE_LOADER, PROP_VALUE_OVERRIDEN_BY_SERVICE_LOADER);
         props.put(PROP_NAME_SAME_ORDINALITY_OVERRIDE, PROP_VALUE_SAME_ORDINALITY_OVERRIDE);
     }
