+++ testsuite/integration/microprofile/src/test/resources/org/wildfly/test/integration/microprofile/config/smallrye/management/config_source/from_class/module_service_loader.xml
